refactor: move UI files to be normal NextJS pages and update generation
