--- examples/demo-viewer/agent/demo/shared_state/style.css
@@ -1,10 +1,0 @@
-.copilotKitWindow {
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-}
-
-.copilotKitHeader {
-  border-top-left-radius: 5px !important;
-  background-color: #fff;
-  color: #000;
-  border-bottom: 0px;
-}
