+++ examples/demo-viewer/src/app/feature/agentic_chat/README.mdx
