+++ examples/demo-viewer/src/app/feature/agentic_chat/page.tsx
@@ -18,7 +18,7 @@ const AgenticChat: React.FC = () => {
 };
 
 const Chat = () => {
-  const [background, setBackground] = useState<string>("#fefefe");
+  const [background, setBackground] = useState<string>("--copilot-kit-background-color");
 
   useCopilotAction({
     name: "change_background",
@@ -44,7 +44,7 @@ const Chat = () => {
     >
       <div className="w-8/10 h-8/10 rounded-lg">
         <CopilotChat
-          className="h-full rounded-lg"
+          className="h-full rounded-2xl"
           labels={{ initial: "Hi, I'm an agent. Want to chat?" }}
         />
       </div>
