+++ examples/demo-viewer/src/app/feature/agentic_chat/style.css
@@ -4,9 +4,9 @@
   border-top-left-radius: 0.75rem;
   border-top-right-radius: 0.75rem;
   border: 1px solid var(--copilot-kit-separator-color) !important;
-  background-color: #fff;
 }
-
+  
 .copilotKitChat {
   background-color: #fff !important;
 }
+
