+++ examples/demo-viewer/src/app/feature/agentic_generative_ui/README.mdx
