+++ examples/demo-viewer/src/app/feature/agentic_generative_ui/page.tsx
@@ -71,11 +71,11 @@ const Chat = () => {
   });
 
   return (
-    <div className="flex justify-center items-center h-screen w-screen">
-      <div className="w-8/10 h-8/10">
+    <div className="flex justify-center items-center h-full w-full">
+      <div className="w-8/10 h-8/10 rounded-lg">
         <CopilotChat
-          className="h-full rounded-lg"
-          labels={{ initial: "Hi, I'm an agent. I can do anything, just ask!" }}
+          className="h-full rounded-2xl"
+          labels={{ initial: "Hi, I'm an agent! I can help you with anything you need and will show you progress as I work. What can I do for you?" }}
         />
       </div>
     </div>
