+++ examples/demo-viewer/src/app/feature/multi_agent_flows/style.css
@@ -127,10 +127,3 @@
 .copilotKitWindow {
   box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
 }
-
-.copilotKitHeader {
-  border-top-left-radius: 5px !important;
-  background-color: #fff;
-  color: #000;
-  border-bottom: 0px;
-}
