+++ examples/demo-viewer/src/app/feature/predictive_state_updates/style.css
@@ -128,9 +128,3 @@
   box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
 }
 
-.copilotKitHeader {
-  border-top-left-radius: 5px !important;
-  background-color: #fff;
-  color: #000;
-  border-bottom: 0px;
-}
