+++ examples/demo-viewer/src/app/feature/shared_state/style.css
@@ -1,7 +1,3 @@
 .copilotKitWindow {
   box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
 }
-
-.copilotKitHeader {
-  border-top-left-radius: 5px !important;
-}
