+++ examples/demo-viewer/src/app/feature/tool_based_generative_ui/README.mdx
