+++ examples/demo-viewer/src/app/feature/tool_based_generative_ui/style.css
@@ -1,0 +1,3 @@
+.copilotKitWindow {
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+}
