+++ examples/demo-viewer/src/components/code-editor/code-editor.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Editor from "@monaco-editor/react";
 import { DemoFile } from "@/types/demo";
-
+import { useTheme } from "next-themes";
 interface CodeEditorProps {
   file?: DemoFile;
   onFileChange?: (fileName: string, content: string) => void;
@@ -13,6 +13,8 @@ export function CodeEditor({ file, onFileChange }: CodeEditorProps) {
       onFileChange(file!.name, value);
     }
   };
+
+  const theme = useTheme();
 
   return file ? (
     <div className="h-full flex flex-col">
@@ -31,7 +33,7 @@ export function CodeEditor({ file, onFileChange }: CodeEditorProps) {
             enabled: false,
           },
         }}
-        theme="vs-dark"
+        theme={theme.theme === "dark" ? "vs-dark" : "vs-light"}
       />
     </div>
   ) : (
